requirements/base.txt
Django==4.0.4
djangorestframework==3.13.1
django-rest-swagger==2.2.0

psycopg2-binary==2.9.3

